template/requirements.txt
anthropic>=0.30.0
openai>=1.30.0
tiktoken>=0.7.0
pyyaml>=6.0
python-dotenv>=1.0.0
